Password encrypt private keys when keyPassword is specified in the request
Closes #21.

Also gets rid of newPemCollection(). That function is not consistently used throughout the codebase, so getting rid of it makes the code more consistent.

Also gets rid of a fromResponse() overload in order to reduce the likelihood of creating a PEMCollection without passing the private key and key password.

Also increases test coverage for PEMCollection to test ECDSA support.