Accept private key from Service Generated CSR mode
When using the Service Generated CSR mode, the private key sent by the server was
not recognized, because `PEMCollection.fromResponse()` didn't support encrypted
private keys. This change adds support for encrypted private keys.

Note: the private key is decrypted in `fromResponse()`, but `pemPrivateKey()`
re-encrypts it locally, so `pemPrivateKey()` doesn't return the original
ciphertext.

Closes #22.