Standardize the authentication process for different VCert clients
The way that the Authetication can be acchieved is not standardized
and that depends completely if the client will be  a VCertClient or
a VCertTknClient or for this last mentioned even if the accessToken
was provided or user&password.

For VCertTknClient exists 2 ways to have the client ready to use.
The first one is when the access token is provided; then it will
required to set it to an Authentication object which will be set to
the Config that will be passed to the VCertTknClient Constructor.
The second one is when the user and password is provided, then firstly
it will be required to create the VCertTknClient object with a Config
object with the Authentication object set and after that those
values will be set to an Authentication object that will be passed as
argument to the VCertTknClient.getAccessToken(Authentication) method.

For VCertClient, no matter if it's type TPP or Cloud, similar to
the second case of VCertTknClient, firstly it will required to create
the VCertClient passing a Config object without the Authentication
object set and then after create the Authentication object setting
into it the required credentials(user&password for TPP and APIKey for
VaaS) in order to call the VCertClient.authenticate(Authentication)
method.

With this refactoring, the sdk provides the following features which
applies for both VCertClient and VCertTknClient:
1. Ability to have the client authenticated at the creation time,
setting the Authentication object to the Config object which is passed
to the constructor of the client.
2. Ability to authenticate the client after it was created. If for
some reason it was not possible to set the Authentication object to
the client when it was created, then the authenticate(Authenticate)
method of the client can be called.