fix: VCertTknClient SDK methods are working again
The VCertTknClient SDK methods left to work since the releases v.0.9.0
until v0.9.2 given a mistake which caused the endpoints used in the
background were the specific endpoints declared for the VCertClient.

Also it was created a bunch of tests to test the VCertTknClient SDK
methods in order to confirm they are working correctly.

Resolves: https://github.com/Venafi/vcert-java/issues/115